backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
assemblyai==0.21.0
google-generativeai==0.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
tf-keras==2.15.0
torch==2.1.0
torchvision==0.16.0
email-validator==2.1.0 